reduce diff
